added gif support and refactored downloader
Added .gif and .gifv support changing the extensions to .mp4
Added a class variable called `self._item` that will make it so
that we don't need to pass the entire reddit instance everytime in
`downloader.py`. Just copy it once during the in `_iterate_items` for
each valid item instead of copying up to 10 times per item on the worst
case like it was before this change.

Also refactored the media fetching function calls for each domain to
it's own function to avoid having too much conditionals on
`_iterate_items` function.