refactored configuration of filename saving format
Before, the user had to give on of two required flags (`--by-user` or
`--by-subreddit`) which, in my opinion, became tedious to write everytime
you wanted to download something.

Now, the user can change the filename saving format by giving the
`--config-save username` to change the filename saving to by-username

The default that comes with the program is by-subreddit, which means
that every file the user download will be saved as
subreddit_uniqueid.ext. Where subreddit will be the subreddit name of
the current post item.

The other option is: `--config-save username` which saved the downloaded
media by username_id.ext